+++ functions/src/admin-match-requests-api/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * add prefix option to Elysia apps for Firebase Hosting compatibility ([d1c25c1](https://github.com/markgoho/doula-cooperative/commit/d1c25c10b28889b39247a5dab7b0adcbaf9533f8))
 * add versioning and changelog infrastructure ([bb61cff](https://github.com/markgoho/doula-cooperative/commit/bb61cff47f85e7f7a5a0cdb8877345183ddd611f))
+* **members:** add self-service email change flow ([1d4f851](https://github.com/markgoho/doula-cooperative/commit/1d4f85116258ad8567db656157bdb56d3d9bff9f))
 * migrate admin match requests functions to Elysia API ([#28](https://github.com/markgoho/doula-cooperative/issues/28)) ([5ef51a3](https://github.com/markgoho/doula-cooperative/commit/5ef51a3f2bcacdc165244d2b125765f26c8236be)), closes [#22](https://github.com/markgoho/doula-cooperative/issues/22) [#22](https://github.com/markgoho/doula-cooperative/issues/22)
 * migrate admin messages functions to Elysia API ([#29](https://github.com/markgoho/doula-cooperative/issues/29)) ([79f7c41](https://github.com/markgoho/doula-cooperative/commit/79f7c41b9f94522f32dbf97f8726ea32cab5e01a)), closes [#23](https://github.com/markgoho/doula-cooperative/issues/23) [#23](https://github.com/markgoho/doula-cooperative/issues/23)
 * notify admin about new member signups ([7ba1d3d](https://github.com/markgoho/doula-cooperative/commit/7ba1d3d5543de45d39cc96a9de28d86fb607d3e3))